requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
azure-ai-translation-text==1.0.1
azure-core==1.31.0
azure-cosmos==4.7.0
azure-functions==1.21.3
azure-identity==1.19.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
colorama==0.4.6
cryptography==43.0.3
distro==1.9.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
isodate==0.7.2
jiter==0.7.0
msal==1.31.0
msal-extensions==1.2.0
openai==1.53.0
portalocker==2.10.1
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.9.0
pywin32==308;platform_system == "Windows"
requests==2.32.3
six==1.16.0
sniffio==1.3.1
tqdm==4.66.6
typing_extensions==4.12.2
urllib3==2.2.3
